Made separate CI for deepseek models.
