+++ .github/workflows/deepseek-tests.yml
@@ -8,6 +8,7 @@ on:
   pull_request:
     branches:
       - main
+  workflow_dispatch:
 
 jobs:
     api-tests:
@@ -49,7 +50,8 @@ jobs:
               run: |
                 python -m pip install --upgrade pip
                 pip install -r requirements_tests.txt
-          
-            - name: Run all API tests
+
+            - name: Run deepseek tests
+              if: github.event_name != 'workflow_dispatch'
               run: |
-                pytest ./api-tests/ -v --ignore=./api-tests/test_api_deepseek.py -v
+                pytest ./api-tests/test_api_deepseek.py -v
