+++ tests/api-tests/test_api_pistachio_rerun_retro.py
@@ -54,49 +54,5 @@ def test_rerun_retro_pistachio_claude_m1p0_success():
     assert ['dependencies', 'steps'] == list(response.json().keys())
 
 
-def test_rerun_retro_pistachio_deepseek_m0p1_success():
-    url = f"{BASE_URL}{ENDPOINTS['rerun_retro']}"
-
-    payload = json.dumps({
-        "smiles": MOLECULE_1,
-        "advanced_model": "False",
-        "advanced_prompt": "True",
-        "llm": DEEPSEEK_FIREWORKS_MODEL,
-        "model_version": PISTACHIO_MODEL})
-
-    headers = {
-        'x-api-key': X_API_KEY,
-        'Content-Type': 'application/json'
-    }
-
-    response = requests.request("POST", url, headers=headers, data=payload)
-
-    assert response.status_code == 200
-    assert isinstance(response.json(), dict)
-    assert ['dependencies', 'steps'] == list(response.json().keys())
-
-
-def test_rerun_retro_pistachio_deepseek_m0p0_success():
-    url = f"{BASE_URL}{ENDPOINTS['rerun_retro']}"
-
-    payload = json.dumps({
-        "smiles": MOLECULE_1,
-        "advanced_model": "False",
-        "advanced_prompt": "False",
-        "llm": DEEPSEEK_FIREWORKS_MODEL,
-        "model_version": PISTACHIO_MODEL})
-
-    headers = {
-        'x-api-key': X_API_KEY,
-        'Content-Type': 'application/json'
-    }
-
-    response = requests.request("POST", url, headers=headers, data=payload)
-
-    assert response.status_code == 200
-    assert isinstance(response.json(), dict)
-    assert ['dependencies', 'steps'] == list(response.json().keys())
-
-
 if __name__ == '__main__':
     pytest.main()
